minor update to accomodate composer daemon test cases

+++ modules/composer/provision/environments/dev.js
@@ -122,7 +122,7 @@ var dev = {
                 "src": true,
                 "level": "debug",
                 "formatter": {
-                    outputMode: 'long'
+                    outputMode: 'short'
                 }
 			},
 			"cors": {
